UIDs are 32-bit, not 64

+++ model/mailbox_message.go
@@ -2,5 +2,5 @@ package model
 
 type MailboxMessage struct {
 	MailboxID string `genji:"index"`
-	MessageID uint64 `genji:"index"`
+	MessageID uint32 `genji:"index"`
 }
